+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 11.4
 Changes in 11.3.3
 -----------------
 - Fix PHP 8.1 fatal error: array_key_exists() arg 2 ($array) must be of type array in ListOutput.fnc.php
+- JS fix get form action on submit, not on load in warehouse.js
 
 Changes in 11.3.2
 -----------------